Preliminary work to make merge with master manageable
